Python Backend/requirements.txt
Flask
SQLAlchemy
PyJWT
Werkzeug
